:rotating_light: Remove client module entirely, RIP GRPC
