--- croncat/src/client/query_handlers/mod.rs
@@ -1,3 +1,0 @@
-pub mod auth_query;
-pub mod bank_query;
-pub mod wasm_query;
