+++ croncat/src/lib.rs
@@ -9,12 +9,11 @@ pub use tokio;
 
 // Our modules
 pub mod channels;
-pub mod client;
 pub mod config;
 pub mod errors;
-pub mod grpc;
 pub mod logging;
 pub mod monitor;
+pub mod rpc;
 pub mod store;
 pub mod streams;
 pub mod system;
