+++ croncat/src/rpc/mod.rs
@@ -2,10 +2,12 @@
 //! Use the [cosmos_sdk_proto](https://crates.io/crates/cosmos-sdk-proto) library to create clients for GRPC node requests.
 //!
 
+pub mod client;
 pub mod querier;
 pub mod service;
 pub mod signer;
 
+pub use client::RpcClient;
 pub use querier::GrpcQuerier;
 pub use service::GrpcClientService;
 pub use signer::GrpcSigner;
