+++ croncatd/src/main.rs
@@ -7,8 +7,8 @@ use croncat::{
     //    client::{BankQueryClient, QueryBank},
     config::Config,
     errors::{eyre, Report},
-    grpc::GrpcClientService,
     logging::{self, error, info},
+    rpc::GrpcClientService,
     store::agent::LocalAgentStorage,
     system,
     tokio,
